test(protocol): move signatureFlow for joinSplit validator into index.js
done to be consistent with test standards
