+++ packages/protocol/test/ACE/validators/joinSplit/index.js
@@ -397,6 +397,169 @@ contract('JoinSplit', (accounts) => {
             expect(decoded[0].publicOwner).to.equal(publicOwner.toLowerCase());
             expect(decoded[0].publicValue).to.equal(kPublic);
         });
+
+        it('should succeed for wrong input note owners', async () => {
+            // Test to confirm no signature validation is performed in the JoinSplit.sol validator
+            const { publicKey } = aztecAccounts[0];
+            const { address: address2 } = aztecAccounts[1];
+
+            const note1Value = 20;
+            const note2Value = 50;
+            const note3Value = 30;
+
+            const note1 = await note.create(publicKey, note1Value, address2);
+            const note2 = await note.create(publicKey, note2Value, address2);
+            const note3 = await note.create(publicKey, note3Value);
+
+            const kPublic = 0;
+            const randomAddress = proofUtils.randomAddress();
+
+            const { proofData, expectedOutput } = joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [note2],
+                outputNotes: [note3, note1],
+                senderAddress: accounts[0],
+                inputNoteOwners: aztecAccounts.slice(4, 5), // incorrect account
+                publicOwner: accounts[3],
+                kPublic,
+                validatorAddress: randomAddress,
+            });
+
+            const opts = {
+                from: accounts[0],
+                gas: 4000000,
+            };
+
+            const publicOwner = accounts[3];
+
+            const result = await joinSplitContract.validateJoinSplit(proofData, accounts[0], constants.CRS, opts);
+
+            const decoded = outputCoder.decodeProofOutputs(`0x${padLeft('0', 64)}${result.slice(2)}`);
+
+            expect(decoded[0].outputNotes[0].gamma.eq(note3.gamma)).to.equal(true);
+            expect(decoded[0].outputNotes[0].sigma.eq(note3.sigma)).to.equal(true);
+            expect(decoded[0].outputNotes[0].noteHash).to.equal(note3.noteHash);
+            expect(decoded[0].outputNotes[0].owner).to.equal(note3.owner.toLowerCase());
+            expect(decoded[0].outputNotes[1].gamma.eq(note1.gamma)).to.equal(true);
+            expect(decoded[0].outputNotes[1].sigma.eq(note1.sigma)).to.equal(true);
+            expect(decoded[0].outputNotes[1].noteHash).to.equal(note1.noteHash);
+            expect(decoded[0].outputNotes[1].owner).to.equal(note1.owner.toLowerCase());
+
+            expect(decoded[0].inputNotes[0].gamma.eq(note2.gamma)).to.equal(true);
+            expect(decoded[0].inputNotes[0].sigma.eq(note2.sigma)).to.equal(true);
+            expect(decoded[0].inputNotes[0].noteHash).to.equal(note2.noteHash);
+            expect(decoded[0].inputNotes[0].owner).to.equal(note2.owner.toLowerCase());
+
+            expect(decoded[0].publicOwner).to.equal(publicOwner.toLowerCase());
+            expect(decoded[0].publicValue).to.equal(kPublic);
+            expect(result).to.equal(expectedOutput);
+            expect(result).to.equal(expectedOutput);
+        });
+
+        it('should succeed for no inputNoteOwners (i.e. no signatures)', async () => {
+            // Test to confirm no signature validation is performed in the JoinSplit.sol validator
+            const { publicKey } = aztecAccounts[0];
+            const { address: address2 } = aztecAccounts[1];
+
+            const note1Value = 20;
+            const note2Value = 50;
+            const note3Value = 30;
+
+            const note1 = await note.create(publicKey, note1Value, address2);
+            const note2 = await note.create(publicKey, note2Value, address2);
+            const note3 = await note.create(publicKey, note3Value);
+
+            const kPublic = 0;
+
+            const randomAddress = proofUtils.randomAddress();
+            const { proofData, expectedOutput } = joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [note2],
+                outputNotes: [note3, note1],
+                senderAddress: accounts[0],
+                inputNoteOwners: [], // incorrect account
+                publicOwner: accounts[3],
+                kPublic,
+                validatorAddress: randomAddress,
+            });
+
+            const opts = {
+                from: accounts[0],
+                gas: 4000000,
+            };
+
+            const result = await joinSplitContract.validateJoinSplit(proofData, accounts[0], constants.CRS, opts);
+            expect(result.signatures).to.equal(undefined);
+            expect(result).to.equal(expectedOutput);
+        });
+
+        it('should succeed for no validatorAddress', async () => {
+            // Test to confirm no signature validation is performed in the JoinSplit.sol validator
+            const { publicKey } = aztecAccounts[0];
+            const { address: address2 } = aztecAccounts[1];
+
+            const note1Value = 20;
+            const note2Value = 50;
+            const note3Value = 30;
+
+            const note1 = await note.create(publicKey, note1Value, address2);
+            const note2 = await note.create(publicKey, note2Value, address2);
+            const note3 = await note.create(publicKey, note3Value);
+
+            const kPublic = 0;
+
+            const { proofData, expectedOutput } = joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [note2],
+                outputNotes: [note3, note1],
+                senderAddress: accounts[0],
+                inputNoteOwners: aztecAccounts.slice(1, 2), // incorrect account
+                publicOwner: accounts[3],
+                kPublic,
+                validatorAddress: [],
+            });
+
+            const opts = {
+                from: accounts[0],
+                gas: 4000000,
+            };
+
+            const result = await joinSplitContract.validateJoinSplit(proofData, accounts[0], constants.CRS, opts);
+            expect(result.signatures).to.equal(undefined);
+            expect(result).to.equal(expectedOutput);
+        });
+
+        it('should succeed for no inputNoteOwners and no validatorAddress', async () => {
+            // Test to confirm no signature validation is performed in the JoinSplit.sol validator
+            const { publicKey } = aztecAccounts[0];
+            const { address: address2 } = aztecAccounts[1];
+
+            const note1Value = 20;
+            const note2Value = 50;
+            const note3Value = 30;
+
+            const note1 = await note.create(publicKey, note1Value, address2);
+            const note2 = await note.create(publicKey, note2Value, address2);
+            const note3 = await note.create(publicKey, note3Value);
+
+            const kPublic = 0;
+
+            const { proofData, expectedOutput } = joinSplit.encodeJoinSplitTransaction({
+                inputNotes: [note2],
+                outputNotes: [note3, note1],
+                senderAddress: accounts[0],
+                inputNoteOwners: [], // incorrect account
+                publicOwner: accounts[3],
+                kPublic,
+                validatorAddress: [],
+            });
+
+            const opts = {
+                from: accounts[0],
+                gas: 4000000,
+            };
+
+            const result = await joinSplitContract.validateJoinSplit(proofData, accounts[0], constants.CRS, opts);
+            expect(result.signatures).to.equal(undefined);
+            expect(result).to.equal(expectedOutput);
+        });
     });
 
     describe('Failure States', () => {
