+++ packages/protocol/test/ERC1724/ZkAsset.js
@@ -31,7 +31,7 @@ const computeDomainHash = (validatorAddress) => {
     return keccak256(`0x${typedData.encodeMessageData(types, 'EIP712Domain', domain)}`);
 };
 
-contract.only('ZkAsset', (accounts) => {
+contract('ZkAsset', (accounts) => {
     let ace;
     let aztecJoinSplit;
     const canAdjustSupply = false;
